requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
psycopg2==2.9.9
python-dotenv==1.0.1
sendgrid==6.11.0
pandas==2.2.1